feat: convert observe_api to single header library
